Prevent symlink when pwd==target path

+++ install.sh
@@ -24,9 +24,11 @@ temp_path="${parentPath}/bashtools"
 echo "[|X] Deciding home for bashtools... ${temp_path}"
 
 #Sym link Tools to path if one doesn't already exist
-if [[ ! -f ${temp_path} ]]; then
+if [[ ! -f ${temp_path} ]] && [[ ${PWD} != ${temp_path} ]]; then
 	ln -s ${PWD}/Tools ${temp_path}
 	echo -e "[|X] Creating symlink from\n\t-> ${temp_path} to ${PWD}/Tools"
+else
+	echo "[|X] No symlink; PWD==target path!"
 fi
 
 #Add lines to .bashrc
